Added step 2 - Simple offer
• The shop decides to introduce two new offers
   • buy one get one free on Apples
    • 3 for the price of 2 on Oranges
• Update your functions & unit tests accordingly

+++ retail/src/main/java/com/epam/retail/model/ItemEntry.java
@@ -4,7 +4,7 @@ public class ItemEntry {
 
     private String itemName;
     double itemRate;
-    double itemQuantity;
+    int itemQuantity;
     double itemTotal;
 
 
@@ -32,15 +32,15 @@ public class ItemEntry {
         this.itemTotal = itemTotal;
     }
 
-    public double getItemQuantity() {
+    public int getItemQuantity() {
         return itemQuantity;
     }
 
-    public void setItemQuantity(double itemQuantity) {
+    public void setItemQuantity(int itemQuantity) {
         this.itemQuantity = itemQuantity;
     }
 
-    public ItemEntry(String itemName, double itemRate, double itemTotal, double itemQuantity) {
+    public ItemEntry(String itemName, double itemRate, double itemTotal, int itemQuantity) {
         this.itemName = itemName;
         this.itemRate = itemRate;
         this.itemTotal = itemTotal;
